blender: Await result of async function
Previously, the script would always believe that Blender is installed on
the host and attempt to launch it. This is because the exists() method
is declared async: calling it returns a coroutine, which as a non-None
object is considered true as a conditional.

Await it, and use that return value to determine whether Blender is
installed.